[KOGITO-7046] Clarifying responsibilities
