+++ kogito-codegen-modules/kogito-codegen-api/src/main/java/org/kie/kogito/codegen/api/context/KogitoBuildContext.java
@@ -160,6 +160,7 @@ public interface KogitoBuildContext extends DroolsModelBuildContext {
                 .collect(Collectors.toUnmodifiableMap(key -> key,
                         key -> getApplicationProperty(key).get()));
     }
+    
 
     interface Builder {
