+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/main/java/org/kie/kogito/serverless/workflow/parser/ServerlessWorkflowParser.java
@@ -20,26 +20,14 @@ import java.io.Reader;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.eclipse.microprofile.openapi.OASFactory;
 import org.eclipse.microprofile.openapi.models.tags.Tag;
-import org.jbpm.process.core.context.variable.VariableScope;
 import org.jbpm.process.core.datatype.impl.type.ObjectDataType;
 import org.jbpm.ruleflow.core.Metadata;
-import org.jbpm.ruleflow.core.RuleFlowNodeContainerFactory;
 import org.jbpm.ruleflow.core.RuleFlowProcessFactory;
-import org.jbpm.ruleflow.core.factory.JoinFactory;
-import org.jbpm.ruleflow.core.factory.NodeFactory;
-import org.jbpm.ruleflow.core.factory.SplitFactory;
-import org.jbpm.ruleflow.core.factory.SubProcessNodeFactory;
-import org.jbpm.ruleflow.core.factory.TimerNodeFactory;
-import org.jbpm.workflow.core.impl.DataAssociation;
-import org.jbpm.workflow.core.impl.DataDefinition;
-import org.jbpm.workflow.core.node.Join;
-import org.jbpm.workflow.core.node.Split;
 import org.kie.kogito.codegen.api.GeneratedInfo;
 import org.kie.kogito.codegen.api.context.KogitoBuildContext;
 import org.kie.kogito.internal.process.runtime.KogitoWorkflowProcess;
@@ -55,11 +43,8 @@ import org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import io.serverlessworkflow.api.Workflow;
-import io.serverlessworkflow.api.events.EventDefinition;
 import io.serverlessworkflow.api.workflow.Constants;
 
-import static org.jbpm.process.core.timer.Timer.TIME_DURATION;
-import static org.jbpm.ruleflow.core.Metadata.UNIQUE_ID;
 import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.processResourceFile;
 
 public class ServerlessWorkflowParser {
@@ -154,86 +139,6 @@ public class ServerlessWorkflowParser {
             processInfo = parseProcess();
         }
         return processInfo;
-    }
-
-    public static <T extends RuleFlowNodeContainerFactory<T, ?>> SubProcessNodeFactory<T> subprocessNode(SubProcessNodeFactory<T> nodeFactory, String inputVar, String outputVar) {
-        Map<String, String> types = Collections.singletonMap(DEFAULT_WORKFLOW_VAR, JSON_NODE);
-        DataAssociation inputDa = new DataAssociation(
-                new DataDefinition(inputVar, inputVar, JSON_NODE),
-                new DataDefinition(DEFAULT_WORKFLOW_VAR, DEFAULT_WORKFLOW_VAR, JSON_NODE), null, null);
-        DataAssociation outputDa = new DataAssociation(
-                new DataDefinition(DEFAULT_WORKFLOW_VAR, DEFAULT_WORKFLOW_VAR, JSON_NODE),
-                new DataDefinition(outputVar, outputVar, JSON_NODE), null, null);
-
-        VariableScope variableScope = new VariableScope();
-        return nodeFactory
-                .independent(true)
-                .metaData("BPMN.InputTypes", types)
-                .metaData("BPMN.OutputTypes", types)
-                .mapDataInputAssociation(inputDa)
-                .mapDataOutputAssociation(outputDa)
-                .context(variableScope)
-                .defaultContext(variableScope);
-    }
-
-    public static <T extends NodeFactory<T, P>, P extends RuleFlowNodeContainerFactory<P, ?>> T sendEventNode(NodeFactory<T, P> actionNode,
-            EventDefinition eventDefinition, String inputVar) {
-        return actionNode
-                .name(eventDefinition.getName())
-                .metaData(Metadata.EVENT_TYPE, "message")
-                .metaData(Metadata.MAPPING_VARIABLE, inputVar)
-                .metaData(Metadata.TRIGGER_REF, eventDefinition.getType())
-                .metaData(Metadata.MESSAGE_TYPE, JSON_NODE)
-                .metaData(Metadata.TRIGGER_TYPE, "ProduceMessage");
-    }
-
-    public static <T extends NodeFactory<T, P>, P extends RuleFlowNodeContainerFactory<P, ?>> T messageNode(T nodeFactory, EventDefinition eventDefinition, String inputVar) {
-        return nodeFactory
-                .name(eventDefinition.getName())
-                .metaData(Metadata.EVENT_TYPE, "message")
-                .metaData(Metadata.TRIGGER_MAPPING, inputVar)
-                .metaData(Metadata.TRIGGER_REF, eventDefinition.getType())
-                .metaData(Metadata.MESSAGE_TYPE, JSON_NODE)
-                .metaData(Metadata.TRIGGER_TYPE, "ConsumeMessage")
-                .metaData(Metadata.DATA_ONLY, isDataOnly(eventDefinition));
-    }
-
-    // TODO remove when SDK is updated to include dataOnly in EventDefinition, see https://github.com/serverlessworkflow/sdk-java/issues/183
-    private static Boolean isDataOnly(EventDefinition eventDefinition) {
-        Boolean result = Boolean.TRUE;
-        Map<String, String> metadata = eventDefinition.getMetadata();
-        final String dataOnlyKey = "dataOnly";
-        if (metadata != null && metadata.containsKey(dataOnlyKey)) {
-            result = Boolean.parseBoolean(metadata.get(dataOnlyKey));
-        }
-        return result;
-    }
-
-    public static <T extends RuleFlowNodeContainerFactory<T, ?>> SplitFactory<T> eventBasedExclusiveSplitNode(SplitFactory<T> nodeFactory) {
-        return nodeFactory.name("ExclusiveSplit_" + nodeFactory.getNode().getId())
-                .type(Split.TYPE_XAND)
-                .metaData(UNIQUE_ID, Long.toString(nodeFactory.getNode().getId()))
-                .metaData("EventBased", "true");
-    }
-
-    public static <T extends RuleFlowNodeContainerFactory<T, ?>> SplitFactory<T> exclusiveSplitNode(SplitFactory<T> nodeFactory) {
-        return nodeFactory.name("ExclusiveSplit_" + nodeFactory.getNode().getId())
-                .type(Split.TYPE_XOR)
-                .metaData(UNIQUE_ID, Long.toString(nodeFactory.getNode().getId()));
-    }
-
-    public static <T extends RuleFlowNodeContainerFactory<T, ?>> JoinFactory<T> joinExclusiveNode(JoinFactory<T> nodeFactory) {
-        return nodeFactory.name("ExclusiveJoin_" + nodeFactory.getNode().getId())
-                .type(Join.TYPE_XOR)
-                .metaData(UNIQUE_ID, Long.toString(nodeFactory.getNode().getId()));
-    }
-
-    public static <T extends RuleFlowNodeContainerFactory<T, ?>> TimerNodeFactory<T> timerNode(TimerNodeFactory<T> nodeFactory, String duration) {
-        return nodeFactory.name("TimerNode_" + nodeFactory.getNode().getId())
-                .type(TIME_DURATION)
-                .delay(duration)
-                .metaData(UNIQUE_ID, Long.toString(nodeFactory.getNode().getId()))
-                .metaData("EventType", "Timer");
     }
 
     private static void loadConstants(Workflow workflow, ParserContext parserContext) {
