+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/main/java/org/kie/kogito/serverless/workflow/parser/handlers/ActionType.java
@@ -15,11 +15,7 @@
  */
 package org.kie.kogito.serverless.workflow.parser.handlers;
 
-import java.util.Map;
-
 import io.serverlessworkflow.api.functions.FunctionDefinition;
-
-import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.isOpenApiOperation;
 
 enum ActionType {
     REST("rest"),
@@ -54,7 +50,7 @@ enum ActionType {
     public static ActionType from(FunctionDefinition actionFunction) {
         switch (actionFunction.getType()) {
             case REST:
-                return isOpenApiOperation(actionFunction) ? ActionType.OPENAPI : fromMetadata(actionFunction.getMetadata());
+                return ActionType.OPENAPI;
             case EXPRESSION:
                 return ActionType.EXPRESSION;
             case CUSTOM:
@@ -75,15 +71,4 @@ enum ActionType {
         throw new UnsupportedOperationException("Unable to recognize custom format " + operation + ", supported custom formats are " + values());
     }
 
-    private static ActionType fromMetadata(Map<String, String> metadata) {
-        String type = metadata != null ? metadata.get("type") : null;
-        if (type != null) {
-            try {
-                return ActionType.valueOf(type.toUpperCase());
-            } catch (IllegalArgumentException ex) {
-                // see return 
-            }
-        }
-        return ActionType.EMPTY;
-    }
 }
