+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/main/java/org/kie/kogito/serverless/workflow/parser/handlers/CompositeContextNodeHandler.java
@@ -15,17 +15,12 @@
  */
 package org.kie.kogito.serverless.workflow.parser.handlers;
 
-import java.io.IOException;
-import java.net.URI;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
-import java.util.function.Supplier;
 
 import org.drools.mvel.java.JavaDialect;
 import org.jbpm.compiler.canonical.descriptors.TaskDescriptor;
@@ -35,7 +30,6 @@ import org.jbpm.ruleflow.core.factory.AbstractCompositeNodeFactory;
 import org.jbpm.ruleflow.core.factory.CompositeContextNodeFactory;
 import org.jbpm.ruleflow.core.factory.NodeFactory;
 import org.jbpm.ruleflow.core.factory.WorkItemNodeFactory;
-import org.kie.kogito.internal.utils.ConversionUtils;
 import org.kie.kogito.jackson.utils.JsonNodeVisitor;
 import org.kie.kogito.jackson.utils.JsonObjectUtils;
 import org.kie.kogito.process.expr.ExpressionHandlerFactory;
@@ -43,36 +37,17 @@ import org.kie.kogito.serverless.workflow.SWFConstants;
 import org.kie.kogito.serverless.workflow.parser.ParserContext;
 import org.kie.kogito.serverless.workflow.parser.ServerlessWorkflowParser;
 import org.kie.kogito.serverless.workflow.parser.SourceFileServerlessWorkflowBindEvent;
-import org.kie.kogito.serverless.workflow.parser.handlers.openapi.OpenAPIDescriptor;
-import org.kie.kogito.serverless.workflow.parser.handlers.openapi.OpenAPIDescriptorFactory;
-import org.kie.kogito.serverless.workflow.suppliers.ApiKeyAuthDecoratorSupplier;
-import org.kie.kogito.serverless.workflow.suppliers.BasicAuthDecoratorSupplier;
-import org.kie.kogito.serverless.workflow.suppliers.BearerTokenAuthDecoratorSupplier;
-import org.kie.kogito.serverless.workflow.suppliers.ClientOAuth2AuthDecoratorSupplier;
-import org.kie.kogito.serverless.workflow.suppliers.CollectionParamsDecoratorSupplier;
-import org.kie.kogito.serverless.workflow.suppliers.ConfigSuppliedWorkItemSupplier;
 import org.kie.kogito.serverless.workflow.suppliers.ExpressionActionSupplier;
 import org.kie.kogito.serverless.workflow.suppliers.ObjectResolverSupplier;
 import org.kie.kogito.serverless.workflow.suppliers.ParamsRestBodyBuilderSupplier;
-import org.kie.kogito.serverless.workflow.suppliers.PasswordOAuth2AuthDecoratorSupplier;
 import org.kie.kogito.serverless.workflow.suppliers.SysoutActionSupplier;
 import org.kie.kogito.serverless.workflow.utils.ExpressionHandlerUtils;
+import org.kie.kogito.serverless.workflow.utils.OpenAPIOperationId;
 import org.kogito.workitem.rest.RestWorkItemHandler;
 import org.kogito.workitem.rest.auth.ApiKeyAuthDecorator;
-import org.kogito.workitem.rest.auth.ApiKeyAuthDecorator.Location;
 import org.kogito.workitem.rest.auth.BearerTokenAuthDecorator;
-import org.kogito.workitem.rest.auth.ClientOAuth2AuthDecorator;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.github.javaparser.ast.body.Parameter;
-import com.github.javaparser.ast.expr.Expression;
-import com.github.javaparser.ast.expr.LambdaExpr;
-import com.github.javaparser.ast.expr.MethodCallExpr;
-import com.github.javaparser.ast.expr.NameExpr;
-import com.github.javaparser.ast.expr.StringLiteralExpr;
-import com.github.javaparser.ast.type.UnknownType;
 
 import io.serverlessworkflow.api.Workflow;
 import io.serverlessworkflow.api.actions.Action;
@@ -82,24 +57,18 @@ import io.serverlessworkflow.api.functions.FunctionDefinition;
 import io.serverlessworkflow.api.functions.FunctionRef;
 import io.serverlessworkflow.api.functions.SubFlowRef;
 import io.serverlessworkflow.api.interfaces.State;
-import io.swagger.parser.OpenAPIParser;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import io.swagger.v3.oas.models.security.SecurityScheme.In;
-import io.swagger.v3.parser.core.models.SwaggerParseResult;
 
-import static org.kie.kogito.internal.utils.ConversionUtils.concatPaths;
-import static org.kie.kogito.serverless.workflow.io.URIContentLoaderFactory.buildLoader;
-import static org.kie.kogito.serverless.workflow.io.URIContentLoaderFactory.readAllBytes;
-import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.OPENAPI_OPERATION_SEPARATOR;
-import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.getServiceName;
+import static org.kie.kogito.serverless.workflow.parser.handlers.NodeFactoryUtils.fillRest;
+import static org.kie.kogito.serverless.workflow.parser.handlers.NodeFactoryUtils.subprocessNode;
+import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.ACCESS_TOKEN;
+import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.API_KEY;
+import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.API_KEY_PREFIX;
+import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.PASSWORD_PROP;
+import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.USER_PROP;
 import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.resolveFunctionMetadata;
-import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.runtimeOpenApi;
 import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.runtimeRestApi;
 
 public abstract class CompositeContextNodeHandler<S extends State> extends StateHandler<S> {
-
-    private static Logger logger = LoggerFactory.getLogger(CompositeContextNodeHandler.class);
 
     private static final String SCRIPT_TYPE_PARAM = "script";
     private static final String SYSOUT_TYPE_PARAM = "message";
@@ -116,11 +85,6 @@ public abstract class CompositeContextNodeHandler<S extends State> extends State
     private static final String LANG_SEPARATOR = ":";
     private static final String METHOD_SEPARATOR = ":";
     private static final String INTFC_SEPARATOR = "::";
-    private static final String USER_PROP = "username";
-    private static final String PASSWORD_PROP = "password";
-    private static final String API_KEY_PREFIX = "api_key_prefix";
-    private static final String API_KEY = "api_key";
-    private static final String ACCESS_TOKEN = "access_token";
 
     protected CompositeContextNodeHandler(S state, Workflow workflow, ParserContext parserContext) {
         super(state, workflow, parserContext);
@@ -184,7 +148,7 @@ public abstract class CompositeContextNodeHandler<S extends State> extends State
             SubFlowRef subFlowRef,
             String inputVar,
             String outputVar) {
-        return ServerlessWorkflowParser.subprocessNode(
+        return subprocessNode(
                 factory.subProcessNode(parserContext.newId()).name(subFlowRef.getWorkflowId()).processId(subFlowRef.getWorkflowId()).waitForCompletion(true),
                 inputVar,
                 outputVar);
@@ -234,26 +198,30 @@ public abstract class CompositeContextNodeHandler<S extends State> extends State
                         .inMapping(inputVar, WORKITEM_PARAM)
                         .outMapping(WORKITEM_PARAM, outputVar), actionFunction, operation, functionRef.getArguments());
             case REST:
-                return addFunctionArgs(addRestParameters(buildRestWorkItem(embeddedSubProcess, actionFunction, inputVar, outputVar), actionFunction, operation), functionRef);
+                return addFunctionArgs(addRestParameters(buildWorkItem(embeddedSubProcess, actionFunction, inputVar, outputVar), actionFunction, operation), functionRef);
             case OPENAPI:
-                return addFunctionArgs(addOpenApiParameters(buildRestWorkItem(embeddedSubProcess, actionFunction, inputVar, outputVar), actionFunction, operation), functionRef);
+                OpenAPIOperationId operationId = OpenAPIOperationId.fromOperation(operation);
+                notifySourceFileCodegenBindListeners(operationId.getUri().toString());
+                return addFunctionArgs(RestOperationHandlerFactory.get(parserContext, operationId).fillWorkItemHandler(buildWorkItem(embeddedSubProcess, actionFunction, inputVar, outputVar), workflow,
+                        actionFunction), functionRef);
             default:
                 return emptyNode(embeddedSubProcess, actionName);
         }
     }
 
-    private WorkItemNodeFactory<?> buildRestWorkItem(RuleFlowNodeContainerFactory<?, ?> embeddedSubProcess,
+    private WorkItemNodeFactory<?> buildWorkItem(RuleFlowNodeContainerFactory<?, ?> embeddedSubProcess,
             FunctionDefinition actionFunction,
             String inputVar,
             String outputVar) {
-        return embeddedSubProcess
-                .workItemNode(parserContext.newId())
-                .name(actionFunction.getName())
-                .metaData(TaskDescriptor.KEY_WORKITEM_TYPE, RestWorkItemHandler.REST_TASK_TYPE)
-                .workName(RestWorkItemHandler.REST_TASK_TYPE)
+        return embeddedSubProcess.workItemNode(parserContext.newId())
                 .inMapping(inputVar, SWFConstants.MODEL_WORKFLOW_VAR)
-                .outMapping(RestWorkItemHandler.RESULT, outputVar)
-                .workParameter(RestWorkItemHandler.BODY_BUILDER, new ParamsRestBodyBuilderSupplier());
+                .outMapping(RestWorkItemHandler.RESULT, outputVar).name(actionFunction.getName());
+    }
+
+    private void notifySourceFileCodegenBindListeners(String uri) {
+        parserContext.getContext()
+                .getSourceFileCodegenBindNotifier()
+                .ifPresent(notifier -> notifier.notify(new SourceFileServerlessWorkflowBindEvent(workflow.getId(), uri)));
     }
 
     private <T extends RuleFlowNodeContainerFactory<T, ?>> WorkItemNodeFactory<T> addFunctionArgs(WorkItemNodeFactory<T> node, FunctionRef functionRef) {
@@ -331,7 +299,7 @@ public abstract class CompositeContextNodeHandler<S extends State> extends State
             method = resolveFunctionMetadata(actionFunction, "method", parserContext.getContext());
         }
 
-        return node.workParameter(RestWorkItemHandler.URL, url)
+        return fillRest(node.workParameter(RestWorkItemHandler.URL, url)
                 .workParameter(RestWorkItemHandler.METHOD, method)
                 .workParameter(RestWorkItemHandler.USER, runtimeRestApi(actionFunction, USER_PROP, parserContext.getContext()))
                 .workParameter(RestWorkItemHandler.PASSWORD, runtimeRestApi(actionFunction, PASSWORD_PROP, parserContext.getContext()))
@@ -340,106 +308,7 @@ public abstract class CompositeContextNodeHandler<S extends State> extends State
                 .workParameter(RestWorkItemHandler.BODY_BUILDER, new ParamsRestBodyBuilderSupplier())
                 .workParameter(BearerTokenAuthDecorator.BEARER_TOKEN, runtimeRestApi(actionFunction, ACCESS_TOKEN, parserContext.getContext()))
                 .workParameter(ApiKeyAuthDecorator.KEY_PREFIX, runtimeRestApi(actionFunction, API_KEY_PREFIX, parserContext.getContext()))
-                .workParameter(ApiKeyAuthDecorator.KEY, runtimeRestApi(actionFunction, API_KEY, parserContext.getContext()));
-    }
-
-    private <T extends RuleFlowNodeContainerFactory<T, ?>> WorkItemNodeFactory<T> addOpenApiParameters(WorkItemNodeFactory<T> node,
-            FunctionDefinition function,
-            String operation) {
-        int indexOf = function.getOperation().indexOf(OPENAPI_OPERATION_SEPARATOR);
-        String uri = operation.substring(0, indexOf);
-        String serviceName = getServiceName(uri);
-        String operationId = operation.substring(indexOf + OPENAPI_OPERATION_SEPARATOR.length());
-        try {
-            // although OpenAPIParser has built in support to load uri, it messes up when using contextclassloader, so using our retrieval apis to get the content
-            SwaggerParseResult result =
-                    new OpenAPIParser().readContents(new String(readAllBytes(buildLoader(URI.create(uri), parserContext.getContext().getClassLoader(), workflow, function.getAuthRef()))), null, null);
-            OpenAPI openAPI = result.getOpenAPI();
-            if (openAPI == null) {
-                throw new IllegalArgumentException("Problem parsing uri " + uri);
-            }
-            logger.debug("OpenAPI parser messages {}", result.getMessages());
-            OpenAPIDescriptor openAPIDescriptor = OpenAPIDescriptorFactory.of(openAPI, operationId);
-            addSecurity(node, openAPIDescriptor, serviceName);
-
-            WorkItemNodeFactory<T> workItemNodeFactory = node.workParameter(RestWorkItemHandler.URL,
-                    runtimeOpenApi(serviceName, "base_path", String.class, OpenAPIDescriptorFactory.getDefaultURL(openAPI, "http://localhost:8080"),
-                            (key, clazz, defaultValue) -> new ConfigSuppliedWorkItemSupplier<String>(key, clazz, defaultValue, calculatedKey -> concatPaths(calculatedKey, openAPIDescriptor.getPath()),
-                                    new LambdaExpr(new Parameter(new UnknownType(), "calculatedKey"),
-                                            new MethodCallExpr(ConversionUtils.class.getCanonicalName() + ".concatPaths")
-                                                    .addArgument(new NameExpr("calculatedKey")).addArgument(new StringLiteralExpr(openAPIDescriptor.getPath()))))))
-                    .workParameter(RestWorkItemHandler.METHOD, openAPIDescriptor.getMethod())
-                    .workParameter(RestWorkItemHandler.PARAMS_DECORATOR, new CollectionParamsDecoratorSupplier(openAPIDescriptor.getHeaderParams(), openAPIDescriptor.getQueryParams()));
-
-            notifySourceFileCodegenBindListeners(uri);
-
-            return workItemNodeFactory;
-        } catch (IOException e) {
-            throw new IllegalArgumentException("Problem retrieving uri " + uri);
-        }
-    }
-
-    private void notifySourceFileCodegenBindListeners(String uri) {
-        parserContext.getContext()
-                .getSourceFileCodegenBindNotifier()
-                .ifPresent(notifier -> notifier.notify(new SourceFileServerlessWorkflowBindEvent(workflow.getId(), uri)));
-    }
-
-    private ApiKeyAuthDecorator.Location from(In in) {
-        switch (in) {
-            case COOKIE:
-                return Location.COOKIE;
-            case HEADER:
-                return Location.HEADER;
-            case QUERY:
-            default:
-                return Location.QUERY;
-        }
-    }
-
-    private void addSecurity(WorkItemNodeFactory<?> node, OpenAPIDescriptor openAPI, String serviceName) {
-        Collection<Supplier<Expression>> authDecorators = new ArrayList<>();
-        for (SecurityScheme scheme : openAPI.getSchemes()) {
-            switch (scheme.getType()) {
-                case APIKEY:
-                    authDecorators.add(new ApiKeyAuthDecoratorSupplier(scheme.getName(), from(scheme.getIn())));
-                    node.workParameter(ApiKeyAuthDecorator.KEY_PREFIX, runtimeOpenApi(serviceName, API_KEY_PREFIX, parserContext.getContext()))
-                            .workParameter(ApiKeyAuthDecorator.KEY, runtimeOpenApi(serviceName, API_KEY, parserContext.getContext()));
-                    break;
-                case HTTP:
-                    if (scheme.getScheme().equals("bearer")) {
-                        authDecorators.add(new BearerTokenAuthDecoratorSupplier());
-                        node.workParameter(RestWorkItemHandler.AUTH_METHOD, new BearerTokenAuthDecorator()).workParameter(BearerTokenAuthDecorator.BEARER_TOKEN,
-                                runtimeOpenApi(serviceName, ACCESS_TOKEN, parserContext.getContext()));
-                    } else if (scheme.getScheme().equals("basic")) {
-                        authDecorators.add(new BasicAuthDecoratorSupplier());
-                        node.workParameter(RestWorkItemHandler.USER, runtimeOpenApi(serviceName, USER_PROP, parserContext.getContext()))
-                                .workParameter(RestWorkItemHandler.PASSWORD, runtimeOpenApi(serviceName, PASSWORD_PROP, parserContext.getContext()));
-                    }
-                    break;
-                case OAUTH2:
-                    // only support client and password credentials
-                    if (scheme.getFlows().getClientCredentials() != null) {
-                        authDecorators.add(new ClientOAuth2AuthDecoratorSupplier(scheme.getFlows().getClientCredentials().getTokenUrl(), scheme.getFlows().getClientCredentials().getRefreshUrl()));
-                        node.workParameter(ClientOAuth2AuthDecorator.CLIENT_ID, runtimeOpenApi(serviceName, "client_id", parserContext.getContext()))
-                                .workParameter(ClientOAuth2AuthDecorator.CLIENT_SECRET, runtimeOpenApi(serviceName, "client_secret", parserContext.getContext()));
-                    } else if (scheme.getFlows().getPassword() != null) {
-                        authDecorators.add(new PasswordOAuth2AuthDecoratorSupplier(scheme.getFlows().getPassword().getTokenUrl(), scheme.getFlows().getPassword().getRefreshUrl()));
-                        node.workParameter(RestWorkItemHandler.USER, runtimeOpenApi(serviceName, USER_PROP, parserContext.getContext()))
-                                .workParameter(RestWorkItemHandler.PASSWORD, runtimeOpenApi(serviceName, PASSWORD_PROP, parserContext.getContext()));
-                    } else if (scheme.getFlows().getAuthorizationCode() != null) {
-                        logger.warn("Unsupported scheme type {} for authorization code flow {}", scheme.getType(), scheme.getFlows().getAuthorizationCode());
-                    } else if (scheme.getFlows().getImplicit() != null) {
-                        logger.warn("Unsupported scheme type {} for implicit flow {}", scheme.getType(), scheme.getFlows().getImplicit());
-                    }
-                    break;
-                default:
-                    logger.warn("Unsupported scheme type {}", scheme.getType());
-            }
-        }
-        if (!authDecorators.isEmpty()) {
-            node.workParameter(RestWorkItemHandler.AUTH_METHOD, authDecorators);
-        }
+                .workParameter(ApiKeyAuthDecorator.KEY, runtimeRestApi(actionFunction, API_KEY, parserContext.getContext())));
     }
 
     private Map<String, Object> functionsToMap(JsonNode jsonNode) {
