+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/main/java/org/kie/kogito/serverless/workflow/utils/OpenAPIOperationId.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.kie.kogito.serverless.workflow.utils;
+
+import java.net.URI;
+import java.nio.file.Path;
+
+public class OpenAPIOperationId {
+
+    static final String OPENAPI_OPERATION_SEPARATOR = "#";
+    private static final String REGEX_NO_EXT = "[.][^.]+$";
+    private static final String ONLY_CHARS = "[^a-z]";
+
+    public static OpenAPIOperationId fromOperation(String operation) {
+        int indexOf = operation.indexOf(OPENAPI_OPERATION_SEPARATOR);
+        return new OpenAPIOperationId(operation.substring(0, indexOf), operation.substring(indexOf + OPENAPI_OPERATION_SEPARATOR.length()));
+    }
+
+    private final URI uri;
+    private final String operationId;
+    private final String fileName;
+
+    private OpenAPIOperationId(String uri, String operationId) {
+        this.uri = URI.create(uri);
+        this.operationId = operationId;
+        this.fileName = Path.of(uri).getFileName().toString().toLowerCase();
+    }
+
+    public URI getUri() {
+        return uri;
+    }
+
+    public String getOperationId() {
+        return operationId;
+    }
+
+    @Override
+    public String toString() {
+        return "OpenApiOperationId [uri=" + uri + ", operationId=" + operationId + "]";
+    }
+
+    public String getServiceName() {
+        return sanitizeName(removeExt(fileName));
+    }
+
+    public String getFileName() {
+        return sanitizeName(fileName);
+    }
+
+    private static String removeExt(String fileName) {
+        return fileName.replaceFirst(REGEX_NO_EXT, "");
+    }
+
+    private static String sanitizeName(String name) {
+        return name.replaceAll(ONLY_CHARS, "");
+    }
+
+}
