+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/java/org/kie/kogito/serverless/workflow/utils/WorkflowUtilsTest.java
@@ -31,7 +31,6 @@ import io.serverlessworkflow.api.mapper.YamlObjectMapper;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.getOpenApiProperty;
-import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.getServiceName;
 import static org.kie.kogito.serverless.workflow.utils.ServerlessWorkflowUtils.resolveFunctionMetadata;
 
 public class WorkflowUtilsTest {
@@ -71,10 +70,4 @@ public class WorkflowUtilsTest {
         assertThat(getOpenApiProperty("testfunction", "base_path", context, String.class, "http://localhost:8080")).isEqualTo("http://localhost:8282");
         assertThat(getOpenApiProperty("testfunction1", "base_path2", context, Integer.class, 0)).isZero();
     }
-
-    @Test
-    public void testGetServiceName() {
-        assertThat(getServiceName("testMethod_1")).isEqualTo("testmethod");
-    }
-
 }
