+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/expression.sw.json
@@ -12,9 +12,8 @@
     },
     {
       "name": "printMessage",
-      "metadata": {
-        "type": "sysout"
-      }
+      "type": "custom",
+      "operation": "sysout"
     }
   ],
   "states": [
