+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/multiple-operations.sw.json
@@ -6,21 +6,18 @@
   "functions": [
     {
       "name": "helloOne",
-      "metadata": {
-        "type": "script"
-      }
+      "type": "custom",
+      "operation": "script"
     },
     {
       "name": "helloTwo",
-      "metadata": {
-        "type": "script"
-      }
+      "type": "custom",
+      "operation": "script"
     },
     {
       "name": "helloThree",
-      "metadata": {
-        "type": "script"
-      }
+      "type": "custom",
+      "operation": "script"
     }
   ],
   "states": [
