+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-eventstate-multi-eventrefs.sw.json
@@ -17,11 +17,10 @@
     }
   ],
   "functions": [
-    {
+     {
       "name": "greetFunction",
-      "metadata": {
-        "type": "script"
-      }
+      "type": "custom",
+      "operation": "script"
     }
   ],
   "states": [
