+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-eventstate-multi-eventrefs.sw.yml
@@ -11,8 +11,8 @@ events:
     type: kafka
 functions:
   - name: greetFunction
-    metadata:
-      type: script
+    type: custom
+    operation: script
 states:
   - name: WaitForGreeting
     type: event
