+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-eventstate.sw.json
@@ -14,9 +14,8 @@
   "functions": [
     {
       "name": "greetFunction",
-      "metadata": {
-        "type": "script"
-      }
+      "type": "custom",
+      "operation": "script"
     }
   ],
   "states": [
