+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-operation-many-functions.sw.json
@@ -6,15 +6,13 @@
   "functions": [
     {
       "name": "helloWorld",
-      "metadata": {
-        "type": "script"
-      }
+      "operation": "script",
+      "type": "custom"
     },
     {
       "name": "goodbyeWorld",
-      "metadata": {
-        "type": "script"
-      }
+      "operation": "script",
+      "type": "custom"
     }
   ],
   "states": [
