+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-operation-many-functions.sw.yml
@@ -3,11 +3,11 @@ name: test-wf
 start: HelloWorld
 functions:
   - name: helloWorld
-    metadata:
-      type: script
+    type: custom
+    operation: script
   - name: goodbyeWorld
-    metadata:
-      type: script
+    type: custom
+    operation: script
 states:
   - name: HelloWorld
     type: operation
