+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-operation-with-delay.sw.yml
@@ -4,8 +4,8 @@ version: '1.0'
 start: HelloWorld
 functions:
   - name: helloWorld
-    metadata:
-      type: script
+    type: custom
+    operation: script
 states:
   - name: HelloWorld
     type: operation
