+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-operation.sw.json
@@ -7,9 +7,8 @@
   "functions": [
     {
       "name": "helloWorld",
-      "metadata": {
-        "type": "script"
-      }
+      "operation": "script",
+      "type": "custom"
     }
   ],
   "states": [
