+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/single-service-operation.sw.json
@@ -7,11 +7,8 @@
   "functions": [
     {
       "name": "helloWorld",
-      "metadata": {
-        "interface": "org.something.other.TestService",
-        "operation": "get",
-        "type": "service"
-      }
+      "type": "custom",
+      "operation": "service:org.something.other.TestService::get"
     }
   ],
   "states": [
