+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/transition-produce-event.sw.json
@@ -13,15 +13,13 @@
   "functions": [
     {
       "name": "helloOne",
-      "metadata": {
-        "type": "script"
-      }
+      "operation": "script",
+      "type": "custom"
     },
     {
       "name": "helloTwo",
-      "metadata": {
-        "type": "script"
-      }
+     "operation": "script",
+      "type": "custom"
     }
   ],
   "states": [
