+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/transition-produce-multi-events.sw.json
@@ -28,15 +28,13 @@
   "functions": [
     {
       "name": "helloOne",
-      "metadata": {
-        "type": "script"
-      }
+      "operation": "script",
+      "type": "custom"
     },
     {
       "name": "helloTwo",
-      "metadata": {
-        "type": "script"
-      }
+      "operation": "script",
+      "type": "custom"
     }
   ],
   "states": [
