+++ kogito-serverless-workflow/kogito-serverless-workflow-builder/src/test/resources/exec/transition-produce-multi-events.sw.yml
@@ -16,11 +16,11 @@ events:
     type: kafka
 functions:
   - name: helloOne
-    metadata:
-      type: script
+    type: custom
+    operation: script
   - name: helloTwo
-    metadata:
-      type: script
+    type: custom
+    operation: script
 states:
   - name: HelloWorld
     type: operation
